refactor(All): move out non version related types
