+++ src/appPlatform.ts
