+++ src/appType.ts
