+++ src/index.ts
@@ -2,3 +2,9 @@ export * from './version-1/'
 export * from './version-2/'
 export * from './version-4/'
 export * from './version-5/'
+export * from './appPlatform'
+export * from './appType'
+export * from './dictionary'
+export * from './language'
+export * from './languagesStaticData'
+export * from './translations'
