+++ src/languagesStaticData.ts
