+++ src/version-4/customContent.ts
@@ -1,4 +1,4 @@
-import { LanguageISO } from './language'
+import { LanguageISO } from '../language'
 
 export enum CustomContentType {
   Word = 'Word',
