+++ src/version-4/index.ts
@@ -7,15 +7,9 @@ export * from './user-update'
 export * from './paddle-webhook'
 export * from './progress'
 export * from './content-report-save'
-export * from './dictionary'
 export * from './login-provider'
-export * from './language'
-export * from './languagesStaticData'
-export * from './translations'
 export * from './deprecated-versions'
 export * from './profile'
 export * from './giftSubscription'
-export * from './appPlatform'
-export * from './appType'
 export * from './customContent'
 export * from './delete-profile'
