+++ src/version-4/settings-sync.ts
@@ -1,4 +1,4 @@
-import { Dictionary } from './dictionary'
+import { Dictionary } from '../dictionary'
 import { PayloadRequestBaseV4 } from './base'
 
 export interface PropertyWithMetadata<T> {
